+++ app/src/main/java/com/mapbox/maps/testapp/examples/markersandcallouts/ViewAnnotationShowcaseActivity.kt
@@ -101,7 +101,6 @@ class ViewAnnotationShowcaseActivity : AppCompatActivity(), OnMapClickListener, 
       else -> super.onOptionsItemSelected(item)
     }
   }
-
   private fun prepareStyle(styleUri: String, bitmap: Bitmap) = style(styleUri) {
     +image(BLUE_ICON_ID, bitmap)
     +geoJsonSource(SOURCE_ID) {